remove cast
svn path=/trunk/openspaces/; revision=38208


Former-commit-id: b81ad99f738e9bdf0d4de5f3025462323eee5772

+++ src/sandbox/src/org/openspaces/enhancer/io/ObjectInputOutput.java
@@ -41,7 +41,7 @@ public class ObjectInputOutput {
     }
 
     public static short readShort(ObjectInput ms) throws IOException {
-        return (short) ms.readShort();
+        return ms.readShort();
     }
 
     public static void writeInt(ObjectOutput ms, int n) throws IOException {
